requirements.txt
absl-py==1.4.0
antlr4-python3-runtime==4.9.3
anyascii==0.3.1
apparmor==3.1.2
appdirs==1.4.4
application-utility==1.3.2
attrs==22.1.0
autocommand==2.2.2
bcrypt==4.0.1
beautifulsoup4==4.11.1
btrfsutil==6.0.2
cachetools==5.2.0
caffeine-ng==4.0.2
certifi==2022.9.24
cffi==1.15.1
chardet==5.1.0
click==8.1.3
cloudpickle==2.2.0
configobj==5.1.0.dev0
contourpy==1.0.6
cryptography==38.0.4
cssselect2==0.7.0
cycler==0.11.0
Cython==0.29.32
docopt==0.6.2
duplicity==1.2.1
ecole==0.8.1
einops==0.6.0
ewmh==0.1.6
fasteners==0.18
filelock==3.8.2
fonttools==4.38.0
future==0.18.2
google-api-core==2.11.0
google-api-python-client==2.70.0
google-auth==2.15.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.6
googleapis-common-protos==1.57.0
grpcio==1.51.1
gufw==22.4.0
gym==0.26.2
gym-notices==0.0.8
httplib2==0.21.0
hydra-core==1.3.1
idna==3.4
inflect==6.0.2
jaraco.context==4.2.0
jaraco.functools==3.5.2
jaraco.text==3.11.0
Jinja2==3.1.2
joblib==1.2.0
kiwisolver==1.4.4
layoutswitcherlib==0.8.33
LibAppArmor==3.1.2
libfdt==1.6.1
lit==14.0.6.dev0
lxml==4.9.2
manjaro-sdk==0.1
Markdown==3.4.1
MarkupSafe==2.1.1
material-color-utilities-python==0.1.5
matplotlib==3.6.2
more-itertools==9.0.0
netsnmp-python==1.0a1
networkx==3.0
nftables==0.1
npyscreen==4.10.5
numpy==1.24.1
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
oauth2client==4.1.3
oauthlib==3.2.2
omegaconf==2.3.0
ordered-set==4.1.0
packaging==21.3
pacman-mirrors==4.23.2
pandas==1.5.2
paramiko==2.11.1
Pillow==9.3.0
ply==3.11
protobuf==3.20.1
psutil==5.9.4
pulsectl==22.3.2
pwquality==1.4.5
pyaml==21.10.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycairo==1.23.0
pycparser==2.21
pydantic==1.10.2
pydot==1.4.2
PyDrive2==1.15.0
PyGObject==3.42.2
PyNaCl==1.4.0
pyOpenSSL==22.1.0
pyparsing==3.0.9
PySCIPOpt==4.2.0
PySide6==6.4.1
python-dateutil==2.8.2
python-xlib==0.31
pytz==2022.6
pyxdg==0.28
PyYAML==6.0
ranger-fm==1.9.3
regex==2022.10.31
reportlab==3.6.12
requests==2.28.1
requests-file==1.5.1
requests-oauthlib==1.3.1
rsa==4.9
scikit-learn==1.2.0
scipy==1.10.0
seaborn==0.12.2
setproctitle==1.3.2
shiboken6==6.4.1
six==1.16.0
smbus==1.1
soupsieve==2.3.2.post1
svglib==1.4.1
systemd-python==235
team==1.0
tensorboard==2.11.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorboardX==2.5.1
threadpoolctl==3.1.0
tinycss2==1.2.1
tldextract==3.4.0
tomli==2.0.1
torch==1.13.1
torch-geometric==2.2.0
torch-scatter==2.1.0
torch-sparse==0.6.16
torch-tb-profiler==0.4.1
tqdm==4.64.1
trove-classifiers==2022.12.22
typing_extensions==4.4.0
ueberzug==18.1.9
ufw==0.36.1
uritemplate==4.1.1
urllib3==1.26.12
validate==5.1.0.dev0
validate-pyproject==0.10.1
webencodings==0.5.1
Werkzeug==2.2.2
Yapsy==1.12.2
